Because implementations could run on 2 byte sized compressed instructions check the instruction through a white list of (at the time roughly) 198 instructions and determine what it is, then set the location offset to 4.
If the instruction is not found assume it is a compressed instruction and
set location offset to 2.  If it is a compressed instruction be verbal about
this in a db_printf(), we don't know which since we don't have a whitelist for
this.

A debug output looks something like this:

panic: Unknown kernel exception 5 trap value 0

Stopped at      0xffffffc0003ae1b6:     jal
    TID    PID    UID     PRFLAGS     PFLAGS  CPU  COMMAND
*     0      0      0     0x10000      0x200    0  swapper
https://www.openbsd.org/ddb.html describes the minimum info required in bug
reports.  Insufficient info makes it difficult to find and fix bugs.
ddb> show panic